Resolved #2428 - CV2: Introduce a way to override survey element (question) adorners
