+++ packages/survey-creator-react/src/adorners/CellQuestion.tsx
