+++ packages/survey-creator-react/src/adorners/QuestionRating.tsx
@@ -21,10 +21,9 @@ export class QuestionRatingAdornerComponent extends QuestionAdornerComponent {
     return this.model as QuestionRatingAdornerViewModel;
   }
 
-  renderElementPlaceholder(): JSX.Element {
-    const question = this.props.question;
+  protected renderFooter(): JSX.Element {
     const model = this.ratingModel;
-    return (
+    return (<>
       <div className="svc-rating-question-controls svc-item-value-controls">
         {model.allowRemove ? attachKey2click(<span
           className="svc-item-value-controls__button svc-item-value-controls__remove"
@@ -39,7 +38,8 @@ export class QuestionRatingAdornerComponent extends QuestionAdornerComponent {
           <SvgIcon size={16} iconName={"icon-add_16x16"}></SvgIcon>
         </span>) : null}
       </div>
-    );
+      {super.renderFooter()}
+    </>);
   }
 }
